+++ packages/std/src/testing/mock.rs
@@ -352,7 +352,7 @@ fn validate_length(bytes: &[u8]) -> StdResult<()> {
 /// Create an env:
 ///
 /// ```
-/// # use cosmwasm_std::{Addr, BlockInfo, ContractInfo, Env, Timestamp, TransactionInfo};
+/// # use cosmwasm_std::{Addr, Binary, BlockInfo, ContractInfo, Env, Timestamp, TransactionInfo};
 /// use cosmwasm_std::testing::mock_env;
 ///
 /// let env = mock_env();
@@ -362,7 +362,7 @@ fn validate_length(bytes: &[u8]) -> StdResult<()> {
 ///         time: Timestamp::from_nanos(1_571_797_419_879_305_533),
 ///         chain_id: "cosmos-testnet-14002".to_string(),
 ///     },
-///     transaction: Some(TransactionInfo { index: 3 }),
+///     transaction: Some(TransactionInfo::new(3, Binary::from_hex("E5469DACEC17CEF8A260FD37675ED87E7FB6A2B5AD95193C51308006C7E494B3").unwrap())),
 ///     contract: ContractInfo {
 ///         address: Addr::unchecked("cosmwasm1jpev2csrppg792t22rn8z8uew8h3sjcpglcd0qv9g8gj8ky922tscp8avs"),
 ///     },
@@ -531,7 +531,13 @@ impl Envs {
                 time,
                 chain_id: self.chain_id.clone(),
             },
-            transaction: Some(TransactionInfo { index: 3 }),
+            transaction: Some(TransactionInfo::new(
+                3,
+                Binary::from_hex(
+                    "E5469DACEC17CEF8A260FD37675ED87E7FB6A2B5AD95193C51308006C7E494B3",
+                )
+                .unwrap(),
+            )),
             contract: ContractInfo {
                 address: self.contract_address.clone(),
             },
@@ -1462,7 +1468,13 @@ mod tests {
                     time: Timestamp::from_nanos(1571797419879305533),
                     chain_id: "cosmos-testnet-14002".to_string()
                 },
-                transaction: Some(TransactionInfo { index: 3 }),
+                transaction: Some(TransactionInfo::new(
+                    3,
+                    Binary::from_hex(
+                        "E5469DACEC17CEF8A260FD37675ED87E7FB6A2B5AD95193C51308006C7E494B3"
+                    )
+                    .unwrap(),
+                )),
                 contract: ContractInfo {
                     address: Addr::unchecked(MOCK_CONTRACT_ADDR)
                 }